Add Halt e fix bug instrucoes

+++ src/venus/tradutor/instrucoes/Instrucao3.java
@@ -24,7 +24,7 @@ public class Instrucao3 {
      * @return String da instrucao pronta para ser escrita
      */
     public String getBin(){
-        int conta = (short)((func << 14) | (rc << 11) | (r << 10) | (dontCare << 8) | off);
+        int conta = ((func << 14) | (rc << 11) | (r << 10) | (dontCare << 8) | off);
         String retorno = Integer.toBinaryString(conta);
         String aux = "";
         for(int i = 0;i < 16 - retorno.length();i++){
